dataController/requirements.txt
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
idna==3.6
Naked==0.1.32
pycparser==2.21
pycryptodome==3.20.0
pymongo==4.6.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
urllib3==2.2.1
